+++ apps/desktop/vitest.config.ts
@@ -1,0 +1,15 @@
+/// <reference types="vitest" />
+import { fileURLToPath } from 'url'
+import { defineConfig } from 'vite'
+
+// https://vitejs.dev/config/
+export default defineConfig(() => ({
+  resolve: {
+    alias: [{ find: '@', replacement: fileURLToPath(new URL('./src', import.meta.url)) }],
+  },
+  test: {
+    environment: 'happy-dom',
+    globals: true,
+    reporters: ['json'],
+  },
+}))