+++ packages/editor/src/markdown-it/lib/common/entities.js
@@ -1,0 +1,6 @@
+// HTML5 entities map: { name -> utf16string }
+//
+
+
+/*eslint quotes:0*/
+module.exports = require('entities/lib/maps/entities.json')