+++ packages/editor/src/transform/markdown-it-list-checkbox.ts
@@ -1,7 +1,7 @@
-import MarkdownIt from "markdown-it"
-import Core from "markdown-it/lib/parser_core"
-import StateCore from "markdown-it/lib/rules_core/state_core"
-import Token from "markdown-it/lib/token"
+import type MarkdownIt from "@/markdown-it"
+import type Core from "@/markdown-it/lib/parser_core"
+import type StateCore from "@/markdown-it/lib/rules_core/state_core"
+import Token from "@/markdown-it/lib/token"
 
 function isBulletListItemToken(t: Token) {
     return t.type === "list_item_open" && ["*", "-"].includes(t.markup)